# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/resultados": {
      "post": {
        "description": "La operacion es idempotente. Reenviar la misma peticion con el mismo clientOperationId devuelve el resultado ya registrado en lugar de crear otro, de modo que un reintento tras una caida de red es seguro.",
        "operationId": "ActivityResultController_registrar",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegistrarResultadoDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Resultado registrado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultadoRespuestaDto"
                }
              }
            }
          },
          "400": {
            "description": "El cuerpo de la peticion no es valido o incumple una regla del dominio."
          },
          "401": {
            "description": "Falta la sesion o el token no es valido."
          },
          "403": {
            "description": "Hay sesion, pero todavia no hay cuenta. Hay que pasar antes por POST /api/cuenta."
          },
          "404": {
            "description": "La operacion solicitada no esta disponible. Se responde asi tanto si no existe como si pertenece a otra persona, para no revelar cual de las dos cosas ocurre."
          },
          "429": {
            "description": "Demasiadas peticiones."
          }
        },
        "security": [
          {
            "sesion": []
          }
        ],
        "summary": "Registrar el resultado de una actividad",
        "tags": [
          "Resultados"
        ]
      }
    },
...
    "/api/diario": {
      "get": {
        "description": "Ordenadas por día y, dentro de cada día, por hora. Sin parámetros, las de hoy en Colombia. Como mucho 366 días por consulta.",
        "operationId": "DiarioController_consultar",
        "parameters": [
          {
            "name": "desde",
            "required": false,
            "in": "query",
            "description": "Primer día del rango, incluido. Si no viene, el mismo que \"hasta\".",
            "schema": {
              "format": "date",
              "example": "2026-09-27",
              "type": "string"
            }
          },
          {
            "name": "hasta",
            "required": false,
            "in": "query",
            "description": "Último día del rango, incluido. Si no viene, hoy en Colombia.",
            "schema": {
              "format": "date",
              "example": "2026-10-03",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EntradaDelDiarioDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Un día mal escrito o un rango no válido."
          },
          "401": {
            "description": "Falta la sesión o el token no es válido."
          },
          "403": {
            "description": "Hay sesión pero todavía no hay cuenta."
          }
        },
        "security": [
          {
            "sesion": []
          }
        ],
        "summary": "Consultar las anotaciones de un rango de días",
        "tags": [
          "Diario"
        ]
      },
      "post": {
        "description": "Idempotente por clientOperationId: un reintento devuelve la anotación ya guardada. Se puede escribir en un día pasado; la anotación conserva la hora real en que se escribió. Si lo escrito trae una señal de riesgo, la respuesta lo indica y trae las líneas de atención; no se guarda ninguna marca.",
        "operationId": "DiarioController_escribir",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EscribirEnElDiarioDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnotacionGuardadaDto"
                }
              }
            }
          },
          "400": {
            "description": "El cuerpo no es válido, la anotación está vacía o el día es futuro."
          },
          "401": {
            "description": "Falta la sesión o el token no es válido."
          },
          "403": {
            "description": "Hay sesión pero todavía no hay cuenta."
          },
          "413": {
            "description": "El cuerpo supera 1 MB."
          }
        },
        "security": [
          {
            "sesion": []
          }
        ],
        "summary": "Escribir una anotación",
        "tags": [
          "Diario"
        ]
      }
    },
    "/api/diario/{id}": {
      "patch": {
        "description": "Lo que no viene se queda como estaba. Pasada una hora desde que se escribió responde 409 EDICION_FUERA_DE_PLAZO, y si otro dispositivo la cambió entretanto, 409 VERSION_DESACTUALIZADA. En los dos casos no se toca nada: el cliente guarda lo suyo como una anotación nueva (ADR 0009). La hora la impone la base de datos, no solo la API.",
        "operationId": "DiarioController_editar",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditarAnotacionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnotacionGuardadaDto"
                }
              }
            }
          },
          "400": {
            "description": "El cuerpo no es válido o no trae nada que cambiar."
          },
          "404": {
            "description": "La anotación no existe o es de otra persona; las dos se responden igual."
          },
          "409": {
            "description": "EDICION_FUERA_DE_PLAZO o VERSION_DESACTUALIZADA. Guardar como anotación nueva."
          }
        },
        "security": [
          {
            "sesion": []
          }
        ],
        "summary": "Corregir una anotación durante su primera hora",
        "tags": [
          "Diario"
        ]
      }
    },
...
  },
  "info": {
    "title": "VSD Health API",
    "description": "API de VSD Health, herramienta de acompanamiento del bienestar emocional y cognitivo.\n\nVSD Health no diagnostica, no formula medicamentos y no reemplaza la atencion de psicologos, medicos ni psiquiatras. Los resultados que devuelve esta API son orientativos.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "sesion": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Token de acceso de Supabase. El navegador lo obtiene al iniciar sesion y lo envia en la cabecera Authorization."
      }
    },
    "schemas": {
      "RegistrarResultadoDto": {
        "type": "object",
        "properties": {
          "activityId": {
            "type": "string",
            "description": "Identificador de la actividad realizada.",
            "format": "uuid",
            "example": "33333333-3333-4333-a333-333333333333"
          },
          "clientOperationId": {
            "type": "string",
            "description": "Identificador de la operacion, generado por el dispositivo. Reenviar el mismo valor en un reintento evita que se duplique el resultado.",
            "format": "uuid",
            "example": "44444444-4444-4444-b444-444444444444"
          },
          "score": {
            "type": "number",
            "description": "Puntaje crudo obtenido, en la escala de la actividad. Se omite en las actividades que no se valoran, como anotar el movimiento del dia, que producen datos y no una calificacion. Que una actividad se valore lo dice su escala y no su tipo: hay bitacoras que si puntuan. El maximo lo declara la actividad, no esta peticion.",
            "example": 8,
            "minimum": 0
          },
          "metadata": {
            "type": "object",
            "description": "Informacion propia del tipo de actividad: las horas de una bitacora de sueno, las respuestas de un registro emocional. No puede traer claves que ya sean campos propios.",
            "example": {
              "horasDormidas": 6.5,
              "despertares": 2
            }
          },
          "completedAt": {
            "type": "string",
            "description": "Momento en que se completo la actividad. No puede estar en el futuro.",
            "format": "date-time",
            "example": "2026-09-14T11:00:00.000Z"
          }
        },
        "required": [
          "activityId",
          "clientOperationId",
          "completedAt"
        ]
      },
      "RecursoDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "titulo": {
            "type": "string",
            "example": "Línea 192, opción 4"
          },
          "descripcion": {
            "type": "string"
          },
          "tipo": {
            "type": "string",
            "example": "contacto",
            "enum": [
              "contacto",
              "lectura",
              "ejercicio"
            ]
          },
          "cobertura": {
            "type": "string",
            "description": "Donde sirve el recurso. La interfaz deberia mostrarlo: un telefono que solo atiende en una ciudad no ayuda a quien esta fuera de ella.",
            "example": "nacional"
          },
          "enlace": {
            "type": "string",
            "format": "uri"
          }
        },
        "required": [
          "id",
          "titulo",
          "tipo"
        ]
      },
      "ResultadoRespuestaDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "activityId": {
            "type": "string",
            "format": "uuid"
          },
          "nivelOrientativo": {
            "type": "string",
            "description": "Nivel orientativo del resultado. Es una orientacion sobre como fue la actividad, nunca un diagnostico. Ausente en las actividades de registro, que no producen puntaje.",
            "enum": [
              "favorable",
              "en_seguimiento",
              "requiere_atencion"
            ]
          },
          "sugiereAcompanamiento": {
            "type": "boolean",
            "description": "Indica si conviene mostrar recursos de apoyo profesional junto al resultado: por el nivel, o porque lo escrito trae una señal de riesgo."
          },
          "lineasDeAtencion": {
            "description": "Las líneas de atención, ordenadas por alcance, cuando sugiereAcompanamiento es cierto. Vacía en cualquier otro caso. Viajan en la misma respuesta para no depender de una segunda petición.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RecursoDto"
            }
          },
          "metadata": {
            "type": "object",
            "description": "Informacion propia del tipo de actividad, tal como la envio el dispositivo. En una actividad de registro es todo su contenido."
          },
          "completedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "activityId",
          "sugiereAcompanamiento",
          "lineasDeAtencion",
          "completedAt"
        ]
      },
...
      "EntradaDelDiarioDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "dia": {
            "type": "string",
            "description": "El día al que pertenece, en el calendario de Colombia.",
            "example": "2026-10-03",
            "format": "date"
          },
          "titulo": {
            "type": "string",
            "nullable": true,
            "maxLength": 120
          },
          "contenido": {
            "type": "object",
            "additionalProperties": true,
            "description": "El documento del editor: un árbol de nodos con \"type\", y opcionalmente \"content\", \"text\", \"marks\" y \"attrs\". Nunca HTML.",
            "example": {
              "type": "doc",
              "content": [
                {
                  "type": "paragraph",
                  "content": [
                    {
                      "type": "text",
                      "text": "Hoy salí a caminar con "
                    },
                    {
                      "type": "text",
                      "text": "calma",
                      "marks": [
                        {
                          "type": "bold"
                        }
                      ]
                    },
                    {
                      "type": "text",
                      "text": "."
                    }
                  ]
                }
              ]
            }
          },
          "adjuntos": {
            "type": "array",
            "items": {
              "type": "object",
              "additionalProperties": true
            },
            "description": "Los diagramas de la anotación: { id, tipo: \"diagrama\", datos }. Los datos son la escena del editor de diagramas, tal cual."
          },
          "version": {
            "type": "number",
            "description": "Sube con cada edición. Hay que enviarla al editar: si no coincide, alguien la cambió desde otro dispositivo.",
            "example": 1
          },
          "creadaEn": {
            "type": "string",
            "format": "date-time",
            "description": "Cuándo se escribió de verdad."
          },
          "editadaEn": {
            "type": "string",
            "format": "date-time"
          },
          "editableHasta": {
            "type": "string",
            "format": "date-time",
            "description": "Hasta cuándo se puede editar: una hora después de escribirla. Después, lo nuevo va en otra anotación."
          }
        },
        "required": [
          "id",
          "dia",
          "titulo",
          "contenido",
          "adjuntos",
          "version",
          "creadaEn",
          "editadaEn",
          "editableHasta"
        ]
      },
...
      "EscribirEnElDiarioDto": {
        "type": "object",
        "properties": {
          "clientOperationId": {
            "type": "string",
            "description": "Identificador de la operación, generado por el dispositivo. Reenviar el mismo valor en un reintento devuelve la anotación ya guardada en lugar de crear otra.",
            "format": "uuid"
          },
          "dia": {
            "type": "string",
            "description": "El día al que pertenece, en el calendario de Colombia. Si no viene, hoy. Puede ser un día pasado, nunca uno futuro.",
            "example": "2026-10-03",
            "format": "date"
          },
          "titulo": {
            "type": "string",
            "maxLength": 120
          },
          "contenido": {
            "type": "object",
            "additionalProperties": true,
            "description": "El documento del editor, con \"type\": \"doc\". Nunca HTML."
          },
          "adjuntos": {
            "type": "array",
            "items": {
              "type": "object",
              "additionalProperties": true
            },
            "description": "Los diagramas: { id, tipo: \"diagrama\", datos }. Como mucho diez."
          }
        },
        "required": [
          "clientOperationId",
          "contenido"
        ]
      },
      "AnotacionGuardadaDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "dia": {
            "type": "string",
            "description": "El día al que pertenece, en el calendario de Colombia.",
            "example": "2026-10-03",
            "format": "date"
          },
          "titulo": {
            "type": "string",
            "nullable": true,
            "maxLength": 120
          },
          "contenido": {
            "type": "object",
            "additionalProperties": true,
            "description": "El documento del editor: un árbol de nodos con \"type\", y opcionalmente \"content\", \"text\", \"marks\" y \"attrs\". Nunca HTML.",
            "example": {
              "type": "doc",
              "content": [
                {
                  "type": "paragraph",
                  "content": [
                    {
                      "type": "text",
                      "text": "Hoy salí a caminar con "
                    },
                    {
                      "type": "text",
                      "text": "calma",
                      "marks": [
                        {
                          "type": "bold"
                        }
                      ]
                    },
                    {
                      "type": "text",
                      "text": "."
                    }
                  ]
                }
              ]
            }
          },
          "adjuntos": {
            "type": "array",
            "items": {
              "type": "object",
              "additionalProperties": true
            },
            "description": "Los diagramas de la anotación: { id, tipo: \"diagrama\", datos }. Los datos son la escena del editor de diagramas, tal cual."
          },
          "version": {
            "type": "number",
            "description": "Sube con cada edición. Hay que enviarla al editar: si no coincide, alguien la cambió desde otro dispositivo.",
            "example": 1
          },
          "creadaEn": {
            "type": "string",
            "format": "date-time",
            "description": "Cuándo se escribió de verdad."
          },
          "editadaEn": {
            "type": "string",
            "format": "date-time"
          },
          "editableHasta": {
            "type": "string",
            "format": "date-time",
            "description": "Hasta cuándo se puede editar: una hora después de escribirla. Después, lo nuevo va en otra anotación."
          },
          "sugiereAcompanamiento": {
            "type": "boolean",
            "description": "Si lo escrito trae una señal de riesgo. Se calcula en cada respuesta y no se guarda con la anotación."
          },
          "lineasDeAtencion": {
            "description": "Las líneas de atención, ordenadas por alcance (lo nacional primero), cuando sugiereAcompanamiento es cierto. Vacía en otro caso.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RecursoDto"
            }
          }
        },
        "required": [
          "id",
          "dia",
          "titulo",
          "contenido",
          "adjuntos",
          "version",
          "creadaEn",
          "editadaEn",
          "editableHasta",
          "sugiereAcompanamiento",
          "lineasDeAtencion"
        ]
      },
      "EditarAnotacionDto": {
        "type": "object",
        "properties": {
          "version": {
            "type": "number",
            "description": "La versión de la anotación cuando se abrió para editar. Si ya no es la actual, la edición se rechaza con 409 VERSION_DESACTUALIZADA.",
            "minimum": 1
          },
          "titulo": {
            "type": "string",
            "maxLength": 120,
            "nullable": true
          },
          "contenido": {
            "type": "object",
            "additionalProperties": true,
            "description": "El documento del editor completo, con \"type\": \"doc\"."
          },
          "adjuntos": {
            "type": "array",
            "nullable": true,
            "items": {
              "type": "object",
              "additionalProperties": true
            }
          }
        },
        "required": [
          "version"
        ]
      },
...
    }
  }
}